Add FolderLinkCard modal's children components
